MagicAudioPlot: plot offset should be a float

+++ General/foleys_MagicJUCEFactories.cpp
@@ -672,7 +672,7 @@ public:
         auto plotLength = int (getProperty (pPlotLength));
         plot.setPlotLength (plotLength);
 
-        auto plotOffset = int (getProperty (pPlotOffset));
+        auto plotOffset = float (getProperty (pPlotOffset));
         plot.setPlotOffset (plotOffset);
     }
